Refactored code to make the hotswap_tool_test.cpp portable
